+++ tests/runpod_smoke/__init__.py
@@ -1,0 +1,6 @@
+"""Internal package for the RunPod image smoke-test runner.
+
+See ../README.md for usage. test_images.py is the entry point and the only
+file users invoke directly; modules in this package are implementation
+details (subject to refactor) and are not part of any public API.
+"""
